chore: merge unit test files
